+++ src/components/search/UseSearch.jsx
@@ -1,18 +1,19 @@
-import { getClassesData } from '../../util/DataBaseRequests';
 import { useState, useEffect } from 'react';
-import { useSearchParams } from 'react-router-dom';
+import { useSearchParams, useNavigate } from 'react-router-dom';
+import { getClassesData } from '../../util/DataBaseRequests';
 
-const useSearch = () => {
+const useSearch = (clearSearchTerm) => {
   const [classes, setClasses] = useState([]);
   const [currentPage, setCurrentPage] = useState(1);
   const [totalPages, setTotalPages] = useState(1);
-  const [search, setSearch] = useState('');
-  const [category, setCategory] = useState('');
   const [searchParams, setSearchParams] = useSearchParams();
-  const [loading, setLoading] = useState(false);
+  const [loading, setLoading] = useState(true);
+  const navigate = useNavigate();
+
+  const searchTerm = searchParams.get('query') || '';
+  const category = searchParams.get('category') || '';
 
   const fetchClasses = async (searchTerm = '', page = 1) => {
-    setLoading(true);
     try {
       const limit = 5;
       const sortBy = 'classTitle';
@@ -33,27 +34,27 @@ const useSearch = () => {
     } catch (error) {
       console.error('Failed to fetch classes:', error.message);
     } finally {
-      setTimeout(() => {
-        setLoading(false);
-      }, 200);
+      setLoading(false);
     }
   };
-  //this is the way it works for both MainPage(DesktopNav) and SearchPage page
 
   useEffect(() => {
-    if (search) {
-      fetchClasses(search, currentPage);
-    }
-  }, [search, currentPage, category]);
-
-  useEffect(() => {
-    fetchClasses(search, currentPage);
-  }, [search, currentPage, category]);
+    fetchClasses(searchTerm, currentPage);
+  }, [searchTerm, currentPage, category]);
 
   const handleSearch = (searchTerm) => {
-    setSearch(searchTerm);
     setSearchParams({ query: searchTerm });
     setCurrentPage(1);
+  };
+
+  const handleSearchSubmit = (event, searchTerm) => {
+    event.preventDefault();
+    if (searchTerm.trim()) {
+      setLoading(true);
+      handleSearch(searchTerm);
+      navigate(`/search?query=${searchTerm}`);
+      clearSearchTerm();
+    }
   };
 
   const handlePageChange = (page) => {
@@ -67,8 +68,8 @@ const useSearch = () => {
     currentPage,
     totalPages,
     handleSearch,
+    handleSearchSubmit,
     handlePageChange,
-    setCategory,
     loading,
   };
 };
